requirements/test.txt
-r prod.txt

pytest
pytest-flask

mock
tox